refactor: mv authimpl from user to common
