+++ common/app/auth.go
@@ -3,10 +3,10 @@ package app
 import (
 	"github.com/gin-gonic/gin"
 
-	cmctl "victorzhou123/vicblog/common/controller"
+	"victorzhou123/vicblog/common/controller"
 	cmdmerror "victorzhou123/vicblog/common/domain/error"
-	cmprimitive "victorzhou123/vicblog/common/domain/primitive"
-	"victorzhou123/vicblog/user/domain/auth"
+	"victorzhou123/vicblog/common/domain/primitive"
+	"victorzhou123/vicblog/common/domain/auth"
 )
 
 const (
@@ -17,7 +17,7 @@ const (
 
 type AuthMiddleware interface {
 	VerifyToken(*gin.Context)
-	GetUser(*gin.Context) (cmprimitive.Username, error)
+	GetUser(*gin.Context) (primitive.Username, error)
 }
 
 type Auth struct {
@@ -27,7 +27,7 @@ type Auth struct {
 func (m *Auth) VerifyToken(ctx *gin.Context) {
 	token := ctx.GetHeader(headerAuthorization)
 	if token == "" {
-		cmctl.SendError(ctx, cmdmerror.New(
+		controller.SendError(ctx, cmdmerror.New(
 			cmdmerror.ErrorCodeTokenNotFound, "",
 		))
 
@@ -38,7 +38,7 @@ func (m *Auth) VerifyToken(ctx *gin.Context) {
 
 	username, err := m.auth.TokenValid(token)
 	if err != nil {
-		cmctl.SendError(ctx, cmdmerror.New(
+		controller.SendError(ctx, cmdmerror.New(
 			cmdmerror.ErrorCodeTokenInvalid, "",
 		))
 
@@ -53,11 +53,11 @@ func (m *Auth) VerifyToken(ctx *gin.Context) {
 	ctx.Next()
 }
 
-func (m *Auth) GetUser(ctx *gin.Context) (cmprimitive.Username, error) {
+func (m *Auth) GetUser(ctx *gin.Context) (primitive.Username, error) {
 	username := ctx.GetHeader(ContextKeyUsername)
 	if username == "" {
 		return nil, cmdmerror.New(cmdmerror.ErrorCodeUserNotFound, "")
 	}
 
-	return cmprimitive.NewUsername(username)
+	return primitive.NewUsername(username)
 }
