+++ common/domain/auth/auth.go
@@ -1,0 +1,12 @@
+package auth
+
+import "victorzhou123/vicblog/common/domain/primitive"
+
+type Payload struct {
+	UserName primitive.Username
+}
+
+type Auth interface {
+	GenToken(*Payload) (string, error)
+	TokenValid(string) (primitive.Username, error)
+}
