+++ main.go
@@ -34,7 +34,7 @@ func main() {
 	log.Init(&cfg.Common.Log, exitSig)
 
 	// mysql
-	if err := mysql.Init(&cfg.Common.Mysql); err != nil {
+	if err := mysql.Init(&cfg.Common.Infra.Mysql); err != nil {
 		log.Warnf("mysql init failed, error: %s", err.Error())
 	}
