--- user/infrastructure/config.go
@@ -1,11 +1,0 @@
-package infrastructure
-
-import "victorzhou123/vicblog/user/infrastructure/authimpl"
-
-type Config struct {
-	Auth authimpl.Config `json:"auth"`
-}
-
-func (cfg *Config) Validate() error {
-	return cfg.Auth.Validate()
-}
